use nbval-lax

+++ smpl/plot.py
@@ -269,7 +269,7 @@ def fit(func, *adata, **kwargs):
         >>> from smpl import functions as f
         >>> from smpl import plot
         >>> param = plot.fit([0,1,2],[0,1,2],f.line)
-        >>> plot.unv(param).round()[0]
+        >>> float(plot.unv(param).round()[0])
         1.0
 
     """
